Move another propagate operation to the top of `expr_infer_with_hint`
Summary:
In D72250256 I tried to group the attribute-narrowing-relevant expressions near
the top of the match because they will need special handling.

I don't plan to handle if expressions (i.e. `x = y if True else z`) in my v0
support, but I just noticed there's already a TODO her about narrowing (which
I renamed to "refinement" in the TODO comment since that's the term we are
generally using in Pyrefly), and I think it makes sense to pull it up to the
top now with that TODO in mind.

Reviewed By: ndmitchell, grievejia

Differential Revision: D72257346

fbshipit-source-id: bd169b2e6ddd7ecfb4433da0e3b149c460e982fc