Fix firefly animations on pyrefly.org by addin --no-minify flag for yarn build
Summary: With the most recent deploy of the Pyrefly website, it seems like the firefly stopped working (see before in test plan). I've always used `yarn start` while testing instead of `yarn build + yarn serve`, and have not seen this issue. It seems like by default `yarn build` minifies the css, which is a potential cause for issues. Passing in the `--no-minify` flag for now to fix the animations. I will also see if there's an alternative way to fix this

Reviewed By: migeed-z, connernilsen

Differential Revision: D72262497

fbshipit-source-id: 5bcacb01d1207d86154021b9eef26061482c553c