Eliminate early returns from `expr_infer_with_hint`
Summary:
We had a few early returns in `expr_infer_with_hint`:
- one of them was exiting from inside a loop when we hit an error we
  knew was non-recoverable at the top-level control flow
- another was just exiting early to save some nesting

I have two objections to this.

---------------

The first is my own general sense of taste: following the rules of structured
programming unless there's a *very* good reason not to has big advantages - and
the advantages are an order of magnitude *bigger*, not smaller, when in a huge
function - because they make it possible to see constrol flow from a summary of
the the code structure rather than by reading every single line of code. I
don't believe that it is worth giving this advantage up, in general, just to
avoid nesting.

This might be a minority opinion, a few years ago there were at least some
cases where I was explicitly told to use early returns in code review because
the nesting was more important than making control flow visibile. But that was
in Python which is statement-oriented (OCaml doesn't even let you - which I
think is telling), whereas with Rust is an expression-oriented language with
pattern matching, which I think makes structured style even more valuable.

I'm somewhat curious what people think, I'm aware my taste isn't shared
by everyone with regard to early returns.

--------------

The second is that I think this is actually causing minor bugs.

First off, it's very clear that `expr_infer_with_hint` is relying on
assigning it's result to `ty` at the top of the massive match expression,
and then at the very bottom calling `self.record_type_trace`. I suspect
that when we wrote that code, we weren't even aware of the early
returns, which are buried inside a 500-line match; it's probably causing
a gap in IDE behavior.

Second, the early return from inside the loop (which is more defensible
in my opinion than the one that's just saving some indentation) is also
causing us to terminate checking the elements of the tuple if we
encounter a bad unpack. So if there's an expression in the tail of the
tuple that fails to check, the errors will never show up. That's
explicitly undesirable for two reasons:
- First off, in a gradual type system they could suppress the unpack
	error if they know it works at runtime, and we shouldn't fail to warn them
	about other code that might crash, that's unsafe.
- Secondly, we're explicitly trying to avoid cases where a type error
  in one part of the code causes unrelated other code to not type check
	becuase we want to minimize the experience where fixing one error
	suddenly causes a bunch of other errors that aren't technically
	downstream of the first error to show up.

Note that the bug tied to this failure to look at the tail is now closed.

--------------

FWIW I think the second point - that we had two early returns and both
of them were bugs in edge cases - is partial evidence that my opinion about
trying to avoid early returns when it's reasonably doable is at least
justifiable.

I'm not arguing we should never have early returns, just that the bar
for them should be pretty high (for example, if we explicitly needed
a branch to *not* hit the tracing logic at the bottom then it would
be justifiable to me... but to save some indentation not so much).

Reviewed By: rchen152

Differential Revision: D72262806

fbshipit-source-id: bd8bce0b6310d3eb6ceb80e7380c16cf06e83123