Add failing testcase for unknown attritbute
Summary:
I was looking at the top 100 errors in our PyTorch dashboard. The top error is "Object of class `str` has no attribute `default`".

I took one of those occurrences and kept reducing the code till I got this testcase. I then tested it on mypy and Pyright.

Pyright rejected this program but the reavel_type was any.

mypy accepted this program.

I am not 100% sure if this example is representitive of every error  of the kind "Object of class `str` has no attribute `default`" because as I reduced the program, mypy started to compain about other parts of the repo. But the reveal_type continued to be `any` on the mypy and pyright's side, which is inconsistent with our behavior.

Was curious on people's thoughts on what we should do for this program.

Reviewed By: rchen152

Differential Revision: D72094957

fbshipit-source-id: 26dd7867a29c4d5ac6cf3aa11c4c8209f9f55a63