Use a custom order for the output json format in tests
Summary:
In the next diff, we will enable interning of callable names. This would lead to a lot of changes to our json output, since the order of callables will change.
To prevent that, this diff introduces a parameter `~sorted` to functions outputting json, and forces a specific order based on target names instead of their interned ID.

Note that the interned ID is currently deterministic, since it is base on the MD5 of the target. In the future, this might change, hence why this diff also introduces `deterministic_compare` functions. This is intended to be future proof.

Reviewed By: tianhan0

Differential Revision: D72235983

fbshipit-source-id: eb161613e1a0164d90034e89bf956fbd0ecd0f90