Replace the call to get_module_and_definition in model parsing
Summary:
Starting with D70297739, we build a map from a target to its define, stored in shared memory.
We still have a few places that call the underlying function `Target.get_module_and_definition` around the codebase.

This diff replaces that call in model parsing. This should help with performance.

Reviewed By: tianhan0

Differential Revision: D72237410

fbshipit-source-id: 81dbd3b8694a3c8ad2982db0c365b3a523015c71