Create new PythonEnvironment struct for version, site package path, and os
Summary:
# This diff

This diff adds a new `PythonEnvironment` struct for parsing and loading Python environment information. This diff:
1. Pulls out `python_version`, `python_platform`, and `site_package_path` into a new config item, which gets flattened when parsing a `ConfigFile`.
2. Adds more getter functions which fall back to the default if even after checking flags, checking the config, and querying the interpreter, there are still unset values.
3. Updates tests and some internal functionality to reflect the new field

# The next few diffs

The next few diffs will add functionality for trying to determine some values for type checking from the system Python installation if not provided by flags or a config instead of falling back to the default. This is a big UX improvement that will match whatever settings are used by the Python interpreter while executing.

This will be done by making these values optional, and only falling  back to Pyrefly's default values if they're still unset when they're needed. We will try to query (and cache) these `PythonEnvironment` results for use in all the configs we load and use them when possible in loaded configs with unset values. On failure, we fall back to Pyrefly's defaults.

## Upcoming diffs
1. Make future `PythonEnvironment` values optional
2. Create a the `PythonEnvironment` (this diff)
3. Add `python_interpreter` to `ConfigFile`
3. Add functionality to query `python_interpreter` for values in the `PythonEnvironment`
4. Add a function to update the `ConfigFile` with a `PythonEnvironment` so the values are set

Reviewed By: ndmitchell

Differential Revision: D72088787

fbshipit-source-id: e54ba49b4e3a0f4fad244d2adb7cabb6381c9066