Use PythonEnvironment interpreter query in ConfigFile
Summary:
# This diff

This diff adds the functionality to override the `PythonEnvironment` in a `ConfigFile` with values we can query an interpreter for.

We can't have this automatically done by `serde`, since if the user is supplying an explicit interpreter, we won't be able to use it at the time we're deserializing the user's config. Therefore, we instead parse `PythonEnvironment` values as `Option`s, and if unset, we then try to query the `PythonEnvironment` for the values we should be using.

# The next few diffs

The next few diffs will add functionality for trying to determine some values for type checking from the system Python installation if not provided by flags or a config instead of falling back to the default. This is a big UX improvement that will match whatever settings are used by the Python interpreter while executing.

This will be done by making these values optional, and only falling  back to Pyrefly's default values if they're still unset when they're needed. We will try to query (and cache) these `PythonEnvironment` results for use in all the configs we load and use them when possible in loaded configs with unset values. On failure, we fall back to Pyrefly's defaults.

## Upcoming diffs
1. Make future `PythonEnvironment` values optional
2. Create a the `PythonEnvironment`
3. Add `python_interpreter` to `ConfigFile`
3. Add functionality to query `python_interpreter` for values in the `PythonEnvironment`
4. Add a function to update the `ConfigFile` with a `PythonEnvironment` so the values are set (this diff)
5. Cache `PythonEnvironment`s

Reviewed By: grievejia

Differential Revision: D72203341

fbshipit-source-id: 57394b4dcc673f55375b900260f5264ce0be9c28