Add FindError::Ignored as error type
Summary:
# This diff

Adds a new `FindError::Ignored` variant, which represents the fact that this module was not found, but we should ignore it and not emit an error.

# This diff stack

This diff adds the same functionality as mypy's [`ignore_missing_imports`](https://mypy.readthedocs.io/en/stable/config_file.html#confval-ignore_missing_imports) configuration option, which tells the type checker to ignore any import errors from the listed modules and replace the module with `typing.Any`.

This is done through diffs with the following:
- Adding a config option to silence the errors, and add that config's value to `LoaderInputs`
- Renaming `FindError::new()` so that it makes more sense
- Adding a new `FindError` variant and `match` logic to ignore the module (this diff)
- Returning the new `FindError` variant from the `Loader`

Reviewed By: grievejia

Differential Revision: D72272616

fbshipit-source-id: 38ee04a4b9b4a4582aa85bf036faef90ce24b537