Absolutize all config paths before loading
Summary:
It looks like there's been an issue with globbing that causes it to give non-deterministic output. My hunches on why this was happening are:
- getting the parent of `PathBuf("pyrefly.toml")` returns something like `PathBuf("")`, which might be weird when used in globbing.
- globbing with relative directories might have incorrect when we're trying to iterate through them with `Globs::resolve_dir`
- the globs library might have weird behavior with relative paths

I'm leaning more toward one of the first ones, since the globs library has been around for a while, but we're also seeing an issue with globbing on windows, so I'm not sure. Switching this to absolute paths (which we're using anyway when auto finding configs) seems to fix it, so let's just stick to that for now.

Reviewed By: grievejia

Differential Revision: D72084930

fbshipit-source-id: 7696535c7b384c56c4d2df3b048965cd66ffcf45