Add a TypeInfo struct to Pyrefly
Summary:
This is where attribute narrowing will live; for now it is a trivial
wrapper around `Type` so that we can do type tetris without functional
changes until all the plumbing we need is in place.

NOTE: I was originally planning to defer putting this up for review until
I had attribute narrowing partially working, but at this stage I'm at least pretty
confident this is a good direction to move in. I've also realized that by the time
I implement anything end-to-end I might need to rebase past several dozen
merge conflicts. So I'm putting it up for review to get the plumbing landed early
as long as there are no objections to the direction the stack is moving in.

Reviewed By: yangdanny97

Differential Revision: D72319609

fbshipit-source-id: 1374b7942b6fd2a3f1450c66840ce76d879bcd36