Make `Answers.record_recursive` take a `Type` as input
Summary:
Once we add attribute narrowing, we're going to have to deal with types that might appear
inside of answers (and therefore be exposed through `Arc`s) that are not actually
`Arc<Type>` but rather `Arc<...something that contains a Type...>`.

The current API boundary between `traits.rs` and `Answers` is too rigid to handle this,
but if we just clone in the caller it will work fine.

Reviewed By: yangdanny97

Differential Revision: D72355561

fbshipit-source-id: 08be6dc52a274f5dc785a17399f770740ab6a5a5