Make the `Expr` binding display better
Summary:
We were debugging Pyrefly getting stuck on a demo codebase,
and we wound up with a dump of the bindings being solved showing
the binding `{}` which is super confusing at first.

I'm 99% sure that what this actually was was `Binding::Expr(...)`
where the expression involved is an empty dictionary, but we
shouldn't have to guess, the display should include a label for
which case we are in.

Reviewed By: rchen152

Differential Revision: D72409502

fbshipit-source-id: 09d1b180c60199cff69245f84ef0f5072fe6f749