Comment out the tests that repro nondeterminism
Summary:
These tests generally repro nondeterministic behavior deterministically, but
we've been seeing occasional failures in `test_nondeterminism_b`; my best guess
is that it's unlikely but possible for the `nondeterministic` module to fully
analyze *before* `main`, so that on occasion the line defining `y` gets
analyzed prior to the one defining `z` even though `main` only depends on `z`
and `x`.

I heasitate to remove the test altogether because it's useful to have a
mostly-reliable repro of this behavior, but we can't be crashing CI
at random, so in this diff I just comment out the version of the test that
seems prone to race conditions.

Reviewed By: yangdanny97

Differential Revision: D72465687

fbshipit-source-id: 71a6d2686298b8435f70c0dfca793ee6176c0ed9