Split the Binding match
Summary:
There are a lot of `Binding` cases that only ever resolve to a bare type - most
of them actually.

Only a relative handful of cases need to deal with attribute narrowing. For the most
basic support, it's really just four:
 - some new Binding kind to actually do the narrowing for attributes
 - Forward needs to forward everything, including attribute narrows
 - Expr needs to handle them so that attribute chains can resolve with narrowing
 - the existing top-level Narrow needs to propagate attribute narrows

Eventually we'll need to expand to a handful of additional Bindings:
 - We'll need Phi or else nested control flow won't work
 - We'll probably need a new binding kind to narrow attributes on assignment
 - We probably want pattern matching to preserve attribute narrowing
 - NameAssign probably should propagate? (I'm fuzzy on how this works, so not certain yet)

But I think the first four cases will be enough to get the basic support in place
(which is important - there's still a bit more plumbing needed to do before any
of this is testable, and making it testable as early as possible is important).

Reviewed By: ndmitchell

Differential Revision: D72398649

fbshipit-source-id: daab070e4b8c1272303b7eee3bb90f6cf4da5345