Rearrange the big Binding match
Summary:
I am preparing for a diff very similar to D72319606 that will split
the `Binding` handling into two parts: a (mostly higher-level) `TypeInfo`
part that has the right type signature to deal with the possibility
of attribute narrowing, and a (mostly lower-level) `Type` part for the
majority of bindings that produce "fresh value" types and therefore
never use attribute narrowing.

Before I make that change, which is nontrivial, it's helpful to move
all of the Bindings that I expect to eventually need to handle
attribute narrowing up to the top of the match; I'll start with splitting
out the bare minimum number of them to implement the most basic attribute
narrowing logic and move more as we expand support.

Reviewed By: ndmitchell

Differential Revision: D72398647

fbshipit-source-id: 429184fe5fff600ea0e71591f1da21b6cfea32b4