Make type tracing logic safer and more concise
Summary:
When `expr_infer_with_hint` was a single big match, it was necessary
to record traces inside of branches that needed to return `TypeInfo`.

But now that we've split it, we can put the record trace logic in
exactly one place, at the top level.

There's only one remaining risk here: as far as I know it is not possible to
make `expr_infer_with_hint` private in such a way that it can only be called by
`expr_infer_type_info_with_hint`. So if someone eventually makes some other
function in the `expr` module try to use `expr_infer_with_hint` directly,
we could get a bug.

That potential bug was actually already live - I'm glad I decided to change
the name becuase it helped me find a few callers; the function now explicitly
has `no_trace` in the name which will hopefully cause maintainers to pause before
using it, and has a doc comment above it saying not to call it directly.

Reviewed By: ndmitchell

Differential Revision: D72462683

fbshipit-source-id: f718d9c644fccd8a0c35d8892940d827a11fe591