Only build higher order call graphs for reachable callables
Summary:
If a callable is not reachable (where the reachability is determined by the
original call graph building), then we can skip building higher order call
graph for it.

One concern is that the original call graphs may miss some reachabilitiy, since
it does not consider the propagation of functions. Consider a callable that is
not reachable from the original call graphs, and has a call graph. Then, one of
the following cases must happen in the final higher order call graphs:
1. The callable is reachable. Then it means it would have been considered as a
new dependency and hence the global fixpoint would build a higher order call
graph for it.
2. The callable is not reachable. Then it is okay to avoid building a higher
order call graph for it.

Reviewed By: arthaud

Differential Revision: D72294825

fbshipit-source-id: c2fb2d747fddb8f2f087788b9ed5ddc1514d66e1