Add ModelQuery constraint any_overriden_base(..)
Summary:
# Context
We currently rely on a python model generator to annotate thrift handlers.
This model generator needs to find all methods that override methods in a given class, that itself extends another class.

We want to get rid of this python model generator because it requires `pyre query`, which might not be supported by pyrefly.

We could create a `ModelQuery` that generates the same model as the python model generator.
However, there is no current way to express what the python model generator currently does.

# Solution
We need to introduce a new constraint `any_overriden_method(inner_constraint)` that matches against methods if the override a method that satisfies `inner_constraint`. This diff implements that.

Reviewed By: tianhan0

Differential Revision: D73115369

fbshipit-source-id: df535a19a92aaf60fca0f7a1a8d35a668d9a6960