Don't propagate taint with missing subtraces
Summary:
See context in D72785183.

It is possible to end up with models that have a tito `T:LocalReturn` while it doesn't have a `T:ExtraTraceSink` sink. Those always mean that the tito was collapse and is not actually a valid path (i.e false positive). We can safely remove the taint in those cases.

Reviewed By: alexkassil

Differential Revision: D73036555

fbshipit-source-id: 324148b76b79e92eee7400b9d4dad2d02322a080