Avoid propagating property targets
Summary:
See context from the previous diff.

Consider this example:
```
class A:
  property
  def foo():  # A.foo = property(get_foo, set_foo)
    ...

def A.foo@decorated():
  return property(A.foo)
```
After this diff, callable `A.foo@decorated` would return nothing, since we no
longer propagate property targets in higher order call graph building. But this
is fine w.r.t. building higher order call graphs because there is actually no
flow of callables here:
1. `property(A.foo)` creates an object that points to the setter and getter.
2. When accessing the property `A().foo`, it means calling the getter, but the
result of the access is not the getter itself (which is the current behavior).
Hence the getter is not propagated.

Reviewed By: arthaud

Differential Revision: D73617777

fbshipit-source-id: dc1cc5a99d13e089c1cc19c160646620ff776178