Avoid dropping non-parameterized targets when parameterized targets are created
Summary:
At a call site, there are various reasons for not creating parameterized
targets (e.g., exceeding the limit of the number of parameterized targets). In
these cases, we should still keep the non-parameterized targets.

The changes in *.models files are reverted by the next diff, which fixes
another bug.

Reviewed By: arthaud

Differential Revision: D73691384

fbshipit-source-id: 5906b5d319c7bc2d426451542e8ae88fb726d713