When lowering down AST nodes to attribute accesses, add the original node as the origin
Summary:
# Context
With pyre being rewritten in Rust, we need to prepare for the migration of Pysa to Pyrefly.
It is likely that Pyrefly won't support type checking arbitrary expressions, which Pysa currently uses.
Instead, it will most likely provide an API to query the type of an expression at a given location.

Unfortunately, Pysa currently relies on lowering down many AST nodes. This means some expressions such as `a + b` are turned into `a.__add__(b)` before being analyzed in the taint analysis - and actually, this is also done during type checking (in pyre1).

Here is a non-exhaustive list of AST lowering:
* Binary operators such `+`, `-`, `*`, `/`, etc. are turned into `__add__`, `__sub__`, etc. calls
* Comparison operators such as `==`, `<=`, etc. are turned into `__eq__`, `__le__`, etc. calls
* Augmented assignments `a += b`, `-=`, `*=`, etc. are turned into `a = a.__iadd__(b)`, etc.
* Subscript operators such as `d[a]` are turned into `d.__getitem__(a)` or `d.__setitem__(a, x)` depending on the context
* Slices such as `1:2` are turned into `slice(1,2,None)`
* `str()` calls such as `str(a)` are turned into `a.__str__()` calls
* `iter()` calls such are turned into `a.__iter__()` or `a.__aiter__` calls
* `next()` and `anext()` calls are turned into `a.__next__()` or `a.__anext__` calls
* `if foo` are turned into `assert(foo)` and `assert(not foo)` during cfg construction
* match statements are turned into if/then/else statements
* `try … except MyException as e` are turned into: `e = …, assert(isinstance(e, MyException))`
* `with a as e` are turned into `e = a.__enter__()` or `e = a.__aenter__()`
* `while foo:` are turned into `assert(foo)` and `assert(not foo)` during cfg construction
* `for a in b:` are turned into `a = b.__iter__().__next__()` calls (or `__aiter__` / `__anext__`)
* `[e for e in l]` are turned into `e = l.__iter__().__next__()` calls

The problem is that this creates AST nodes that don't exist in the original code, and with the same location as existing AST nodes. For instance, `a + b` is turned into `a.__add__(b)`, and `a.__add__` has the same location has `a` (this would obviously be invalid if returned by the parser).
This will probably lead to bugs when we query pyrefly for the type of `a.__add__`, it will most likely answer with the type of `a`.

# Solution
We will probably need to find a way for pyrefly to return extra information about AST nodes. For instance, query for `a + b` could return the go-to-definition for `a.__add__`.

On the Pysa side, we would need know that `a.__add__` is an "artificial" node that comes from `a + b` so we can properly query Pyrefly.

To do this, we plan to add an `origin` field to fabricated (or "artificial") AST nodes, so we know the node that originated them.

This diff adds the `origin` field to attribute accesses, since the is the most common artificial AST node.

Reviewed By: tianhan0

Differential Revision: D73845705

fbshipit-source-id: 3cbcb6d27bc5c2e2aeb963fe254e5fbb65f642d8